requirements.txt
# SwarmBrain Dependencies

# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Federated Learning
flwr>=1.7.0  # Flower framework
# nvidia-flare>=2.4.0  # Alternative: NVIDIA FLARE (uncomment if using)

# ROS 2 Python bindings (install via apt/rosdep, not pip)
# rclpy
# std_msgs
# geometry_msgs

# Networking and async
aiohttp>=3.9.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
websockets>=12.0

# Data processing
pandas>=2.0.0
opencv-python>=4.8.0
pillow>=10.0.0

# Cryptography
cryptography>=41.0.0
pycryptodome>=3.19.0
# openfhe  # Install separately (see crypto/fhe/openfhe_integration.md)

# Graph processing (for task graphs)
networkx>=3.0

# API and web frameworks
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Monitoring and logging
prometheus-client>=0.19.0
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
structlog>=24.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
Click>=8.1.0
tqdm>=4.66.0

# Database (for mission/task storage)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Message broker clients
pika>=1.3.0  # RabbitMQ
# kafka-python>=2.0.0  # Apache Kafka (uncomment if using)

# Industrial data integration
asyncua>=1.0.0  # OPC UA client for SCADA/PLC
paho-mqtt>=1.6.0  # MQTT client for IoT sensors
requests>=2.31.0  # HTTP client for MES/ERP APIs
influxdb-client>=1.38.0  # InfluxDB for time-series data
# pymodbus>=3.5.0  # Modbus TCP/RTU (uncomment if using)

# Development tools
ipython>=8.20.0
jupyter>=1.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.5.0
mkdocstrings[python]>=0.24.0
